rename all subcomponents main file into index.js
