+++ src/listView/macOs/separator/index.js
