+++ src/masterDetailsView/windows/details/index.js
