+++ src/masterDetailsView/windows/item/index.js
