+++ src/masterDetailsView/windows/master/index.js
