+++ src/navPane/windows/item/content/index.js
