+++ src/navPane/windows/item/index.js
@@ -4,8 +4,8 @@ import Padding, { paddingPropTypes } from '../../../style/padding';
 import Margin, { marginPropTypes } from '../../../style/margin';
 import Background, { backgroundPropTypes } from '../../../style/background/windows';
 import Alignment, { alignmentPropTypes } from '../../../style/alignment';
-import Title from './title/title';
-import Content from './content/content';
+import Title from './title';
+import Content from './content';
 import styles from '../style/windows10';
 import { StyleRoot } from 'radium';
 import { ColorContext, colorPropTypes } from '../../../style/color/windows';
