+++ src/navPane/windows/item/title/index.js
