+++ src/navPane/windows/pane/button/index.js
