+++ src/navPane/windows/pane/item/index.js
