+++ src/segmentedControl/macOs/item/index.js
