+++ src/segmentedControl/macOs/tabs/index.js
